Tests to user resources. Missing delete test

Approved-by: rt <rgtzths@protonmail.com>

+++ CloudPlaces/src/main/java/cloudplaces/webapp/database_queries/UserQueries.java
@@ -60,10 +60,10 @@ public class UserQueries {
     return null;
   }
   
-  public List<House> getWishlist(String userEmail){
+  public Object getWishlist(String userEmail){
     Optional<User> user = userRepo.findById(userEmail);
     if (user.isPresent()){
-      List<Wishlist> wishList = em.createQuery("select u.wishes from User u WHERE u.email = " + userEmail).getResultList();
+      List<Wishlist> wishList = em.createQuery("select u.wishes from User u WHERE u.email = '" + userEmail + "'").getResultList();
       List<House> houses = new ArrayList<>();
       List<House> houseWishes = em.createQuery("select h from House h").getResultList();
       for (Wishlist w : wishList){
@@ -77,7 +77,7 @@ public class UserQueries {
       }
       return houses;
     }
-    return new ArrayList<>();
+    return null;
   }
   
   public Wishlist addToWishlist(String userEmail, long propertyId){
@@ -104,14 +104,14 @@ public class UserQueries {
   public boolean deleteFromWishlist(String email, long property_id) {
     Optional<User> user = userRepo.findById(email);
     Optional<House> house = propertyRepo.findById(property_id);
-    if (user.isPresent()){
+    if (user.isPresent() && house.isPresent()){
       User u = user.get();
       House h = house.get();
       List<Wishlist> userWishlist = u.getWishes();
       List<Wishlist> houseWishlist = h.getWishes();
       for (Wishlist w1 : userWishlist){
         for (Wishlist w2 : houseWishlist){
-          if (w1.getWhishListId() == w2.getWhishListId()){
+          if (w1.getWhishListId().equals(w2.getWhishListId())){
             Optional<Wishlist> wishlist = wishlistRepo.findById(w1.getWhishListId());
             if (wishlist.isPresent()){
               wishlistRepo.deleteById(w1.getWhishListId());
